Update doctest to release 2.5.2 and fixed asserts/requires where unsigned and signed integrals where compared

+++ src/tests/gdlib/gmsobjtests.cpp
@@ -66,7 +66,7 @@ TEST_CASE( "Simple use TBooleanBitArray" )
    bba.SetBit( 4, true );
    REQUIRE_EQ( bba.MemoryUsed(), oldMem );
    auto vec = asBoolVec( bba );
-   REQUIRE_EQ( 5, vec.size() );
+   REQUIRE_EQ( 5u, vec.size() );
    std::vector<bool> expectedVec { false, false, false, true, true };
    for( int i {}; i < 5; i++ )
       REQUIRE_EQ( expectedVec[i], vec[i] );
@@ -89,7 +89,7 @@ TEST_CASE( "Simple use of TXList" )
       nums[i] = i + 1;
       REQUIRE_EQ( i, lst.Add( &nums[i] ) );
    }
-   REQUIRE_EQ( nums.size(), lst.GetCount() );
+   REQUIRE_EQ( static_cast<int>(nums.size()), lst.GetCount() );
    REQUIRE( lst.MemoryUsed() > 0 );
    for( int i = 0; i < lst.GetCount(); i++ )
       REQUIRE_EQ( &nums[i], lst[i] );
